Refactor out mongo forum from ingestor forum
